RecorridoApp/requirements.txt
Flask==3.0.3
flask-marshmallow==1.2.1
flask-cors==4.0.0
marshmallow==3.23
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
marshmallow-sqlalchemy==1.1.0
Flask-Migrate==4.0.7
psycopg2-binary==2.9.10
python-dotenv==1.0.1
